+++ src/stream.rs
@@ -1,5 +1,3 @@
-use nom::Err;
-
 use metadata;
 use frame;
 use subframe;
@@ -22,10 +20,16 @@ pub struct Stream<P: StreamProducer> {
   output: Vec<i32>,
 }
 
+/// Alias for a FLAC stream produced from `Read`.
+pub type StreamReader<R>  = Stream<ReadStream<R>>;
+
+/// Alias for a FLAC stream produced from a byte stream buffer.
+pub type StreamBuffer<'a> = Stream<ByteStream<'a>>;
+
 impl<P> Stream<P> where P: StreamProducer {
   /// Constructor for the default state of a FLAC stream.
   #[inline]
-  pub fn new<R: io::Read>(reader: R) -> io::Result<Stream<ReadStream<R>>> {
+  pub fn new<R: io::Read>(reader: R) -> io::Result<StreamReader<R>> {
     let producer  = ReadStream::new(reader);
     let error_str = "parser: couldn't parse the reader";
 
@@ -56,7 +60,7 @@ impl<P> Stream<P> where P: StreamProducer {
   /// * `ErrorKind::InvalidData` is returned when the data within the file
   ///   isn't valid FLAC data.
   #[inline]
-  pub fn from_file(filename: &str) -> io::Result<Stream<ReadStream<File>>> {
+  pub fn from_file(filename: &str) -> io::Result<StreamReader<File>> {
     File::open(filename).and_then(|file| {
       let producer  = ReadStream::new(file);
       let error_str = format!("parser: couldn't parse the given file {}",
@@ -75,7 +79,7 @@ impl<P> Stream<P> where P: StreamProducer {
   /// * `ErrorKind::InvalidData` is returned when the data within the buffer
   ///   isn't valid FLAC data.
   #[inline]
-  pub fn from_buffer(buffer: &[u8]) -> io::Result<Stream<ByteStream>> {
+  pub fn from_buffer(buffer: &[u8]) -> io::Result<StreamBuffer> {
     let producer  = ByteStream::new(buffer);
     let error_str = "parser: couldn't parse the buffer";